+++ OpenSim/Region/ScriptEngine/Interfaces/IScript.cs
@@ -36,7 +36,7 @@ namespace OpenSim.Region.ScriptEngine.Shared.ScriptBase
         string[] GetApis();
         void InitApi(string name, IScriptApi data);
 
-        int GetStateEventFlags(string state);
+        ulong GetStateEventFlags(string state);
         void ExecuteEvent(string state, string FunctionName, object[] args);
         Dictionary<string,Object> GetVars();
         void SetVars(Dictionary<string,Object> vars);
